requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
fake-useragent==2.2.0
idna==3.11
pyexecjs==1.5.1
requests==2.32.5
six==1.17.0
urllib3==2.5.0
